book_details_fastapi_integration_mongo/requirement.txt
uvicorn~=0.22.0
fastapi~=0.95.1
pymongo~=4.3.3
pydantic~=1.10.7
python-dotenv~=1.0.0
